requirements.txt
python-telegram-bot>=13
pytz
beautifulsoup4
lxml
